Revert "af_lavrresample: don't drop sl/sr channels for 7.1 on ALSA"
This reverts commit 49d8100c724dd59b1dc9388533835306194128de.

Testing shows the channel pairs must indeed be swapped (details see
commit message of the reverted commit). Making the downmix code move
sl/sr to sdl/sdr is not an appropriate solution anymore, and it's
better to fix the unusual channel layout in ao_alsa.c directly.

(Not reverting the change in chmap.c; this is still correct.)